+++ addons/web/controllers/__init__.py
@@ -1,6 +1,16 @@
-# -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import main
+from . import utils
+from . import action
+from . import binary
+from . import database
+from . import dataset
+from . import export
+from . import home
 from . import pivot
 from . import profiling
+from . import report
+from . import session
+from . import view
+from . import webclient
+
+from . import main  # deprecated
